+++ projects/kypo-training-agenda/instance-detection-event-detail/src/components/training-instance-detection-event-detail-component.module.ts
@@ -11,6 +11,12 @@ import { DetectionEventParticipantConcreteService } from '../services/participan
 import { SentinelPipesModule } from '@sentinel/common/pipes';
 import { MatCardModule } from '@angular/material/card';
 import { TrainingInstanceDetectionEventDetailMaterialModule } from './training-instance-detection-event-detail-material.module';
+import { MatTabsModule } from '@angular/material/tabs';
+import { TimelineModule } from '@muni-kypo-crp/command-visualizations/timeline';
+import { DetectionEventForbiddenCommandsService } from '../services/forbidden-commands/detection-event-forbidden-commands.service';
+import { DetectionEventForbiddenCommandsConcreteService } from '../services/forbidden-commands/detection-event-forbidden-commands-concrete.service';
+import { MatExpansionModule } from '@angular/material/expansion';
+import { MatDividerModule } from '@angular/material/divider';
 
 @NgModule({
   imports: [
@@ -19,6 +25,10 @@ import { TrainingInstanceDetectionEventDetailMaterialModule } from './training-i
     SentinelPipesModule,
     TrainingInstanceDetectionEventDetailMaterialModule,
     MatCardModule,
+    MatTabsModule,
+    TimelineModule,
+    MatExpansionModule,
+    MatDividerModule,
   ],
   declarations: [TrainingInstanceDetectionEventDetailComponent],
   providers: [
@@ -27,6 +37,7 @@ import { TrainingInstanceDetectionEventDetailMaterialModule } from './training-i
     { provide: TrainingNavigator, useClass: TrainingDefaultNavigator },
     { provide: DetectionEventService, useClass: DetectionEventConcreteService },
     { provide: DetectionEventParticipantService, useClass: DetectionEventParticipantConcreteService },
+    { provide: DetectionEventForbiddenCommandsService, useClass: DetectionEventForbiddenCommandsConcreteService },
   ],
 })
 export class TrainingInstanceDetectionEventDetailComponentsModule {
@@ -35,7 +46,10 @@ export class TrainingInstanceDetectionEventDetailComponentsModule {
   ): ModuleWithProviders<TrainingInstanceDetectionEventDetailComponentsModule> {
     return {
       ngModule: TrainingInstanceDetectionEventDetailComponentsModule,
-      providers: [{ provide: TrainingAgendaConfig, useValue: config }],
+      providers: [
+        TimelineModule.forRoot(config.visualizationConfig).providers,
+        { provide: TrainingAgendaConfig, useValue: config },
+      ],
     };
   }
 }