+++ projects/kypo-training-agenda/instance-detection-event-detail/src/model/detection-event-forbidden-commands-row-adapter.ts
@@ -1,0 +1,9 @@
+import { DetectedForbiddenCommand, DetectionEventParticipant } from '@muni-kypo-crp/training-model';
+
+/**
+ * Class representing row of cheating detection table
+ */
+export class DetectionEventForbiddenCommandsRowAdapter extends DetectedForbiddenCommand {
+  occurredAtFormatted: string;
+  typeFormatted: string;
+}